Fix problem on History parser on database

+++ app/src/main/java/com/example/MagicShop/model/HistoryRecord.java
@@ -5,18 +5,18 @@ import java.util.Date;
 public class HistoryRecord {
 
     private long product_id;
-    private long price;
+    private float price;
     private UserHistory user;
     private long date;
 
-    public HistoryRecord(long product_id, long price, long date, UserHistory user){
+    public HistoryRecord(long product_id, float price, long date, UserHistory user){
         this.product_id = product_id;
         this.price = price;
         this.user = user;
         this.date = date;
     }
 
-    public static HistoryRecord create(long product_id, long price, long date, UserHistory user){
+    public static HistoryRecord create(long product_id, float price, long date, UserHistory user){
         return new HistoryRecord(product_id, price,date, user);
     }
 
@@ -28,11 +28,11 @@ public class HistoryRecord {
         this.product_id = product_id;
     }
 
-    public long getPrice() {
+    public float getPrice() {
         return price;
     }
 
-    public void setPrice(long price) {
+    public void setPrice(float price) {
         this.price = price;
     }
